Expose function for creating a mongo session without middleware
